backend/pyproject.toml
[project]
name = "video-interview-backend"
version = "0.1.0"
description = "Local-first video interview analysis API"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.110.0",
  "uvicorn[standard]>=0.27.0",
  "python-multipart>=0.0.9",
  "pydantic>=2.6.0",
  "pydantic-settings>=2.2.1",
  "python-dotenv>=1.0.1",
  "google-genai>=0.5.0",
  "rich>=13.7.1",
]

[tool.uv]
dev-dependencies = [
  "ruff>=0.5.7",
]
